Implemented a breadcrumb scheme.
The breadcrumb system is controlled through a new BaseViewModel class. This class has a Crumbs field that you can add objects to that will render based on the XAML in the HomePage. The commands specified on the Crumbs objects will be called as they are tapped or clicked by the user.

This also includes a change to the project type for the mobile view models. Originally, it was set as a normal PCL, however certain windows specific things required it be a UWP class library. So now it is a UWP library.